+++ app/routes/components/IconButton/changelog.codegen.ts
@@ -6,4 +6,5 @@ export const changelog = ''
   + '- 2025-09-19: Support the prop `size` (`sm`, `md` and `lg`).\n'
   + '- 2025-09-26: Use theme tokens for styling.\n'
   + '- 2025-10-11: Code formatting improvements.\n'
+  + '- 2025-10-30: Add the default icon size for SVG children.\n'
   ;