+++ src/components/IconButton/index.ts
@@ -1,4 +1,4 @@
-// IconButton from copy-ui @ 2025-09-26
+// IconButton from copy-ui @ 2025-10-30.
 
 export type { IconButtonProps } from "./IconButton";
 export { IconButton } from "./IconButton";