+++ src/components/IconButton/CHANGELOG.md
@@ -2,3 +2,4 @@
 - 2025-09-19: Support the prop `size` (`sm`, `md` and `lg`).
 - 2025-09-26: Use theme tokens for styling.
 - 2025-10-11: Code formatting improvements.
+- 2025-10-30: Add the default icon size for SVG children.